Revert "unit tests for initialization of database and the database models"
