--- .nyc_output/977eb044f17cbbfe73207aa494711d0f.json
@@ -1,1 +1,0 @@
-{}
