--- .nyc_output/d36fbd41dd0060cf317b669c26442480.json
@@ -1,1 +1,0 @@
-{}
