+++ src/config/config.json
@@ -25,10 +25,10 @@
   },
   "development": {
     "username": "root",
-    "password": "home",
-    "database": "interopdb",
+    "password": "Naomi@123",
+    "database": "interopdb_kibera",
     "host": "localhost",
-    "port": 3306,
+    "port": 3308,
     "dialect": "mysql",
     "logging": false,
     "operatorsAliases": false
